GCD and LCM Program Using the Concept of Recursion
The greatest common divisor (GCD) of two or more numbers is the greatest common factor number that divides them, exactly.
The least common multiple (LCM) is the smallest number that two or more numbers can divide into evenly. To find the LCM, you can use the prime factorization method or list the multiples of each number.