Completed the ISC Banner Program in Java
The names of the teams participating in a competition should be displayed on a banner vertically,  to accommodate as many teams as possible in a single banner. Design a program to accept the names of N teams, where 2<N<9 and display them in vertical order, side by side with a horizontal tab (i.e. eight spaces).