Program to sort each row of Matrix | Java
Write a program to declare a matrix A[ ] [ ] of order (MN) where ‘M’ is the number of 
rows and ‘N’ is the number of columns such that the values of both ‘M’ and ‘N’ must be 
greater than 2 and less than 10. Allow the user to input integers into this matrix. 
Perform the following tasks on the matrix: 
(a) Display the original matrix. 
(b) Sort each row of the matrix in ascending order using any standard sorting 
technique. 
(c) Display the changed matrix after sorting each row.